refactor: split hook injector SDK context lookup
